Improve SidebarBeforeOutput and remove BaseTemplateToolbox
BaseTemplateToolbox was deprecated in 1.35 so remove it, and it's
related tests.

Update tests to use SidebarBeforeOutput hook and add a condition
in the hook to check for isArticle()
