+++ src/MediaWiki/Hooks/SidebarBeforeOutput.php
@@ -19,7 +19,7 @@ use SMW\OptionsAwareTrait;
  * @author StarHeartHunt
  */
 class SidebarBeforeOutput implements HookListener {
-    
+
     use OptionsAwareTrait;
 
 	/**
@@ -47,26 +47,25 @@ class SidebarBeforeOutput implements HookListener {
 		$title = $skin->getTitle();
 
 		if ( $this->canProcess( $title, $skin ) ) {
-			$this->performUpdate( $title, $sidebar );
+			$this->performUpdate( $title, $skin, $sidebar );
 		}
 
 		return true;
 	}
 
-	private function canProcess( Title $title, $skin ) {
+	private function canProcess( Title $title, Skin $skin ) {
 		if ( $title->isSpecialPage() || !$this->namespaceExaminer->isSemanticEnabled( $title->getNamespace() ) ) {
 			return false;
 		}
-		
-		# || !$skin->data['isarticle'] 
-		if ( !$this->isFlagSet( 'smwgBrowseFeatures', SMW_BROWSE_TLINK )) {
+
+		if ( !$skin->getOutput()->isArticle() || !$this->isFlagSet( 'smwgBrowseFeatures', SMW_BROWSE_TLINK ) ) {
 			return false;
 		}
 
 		return true;
 	}
 
-	private function performUpdate( $title, &$sidebar ) {
+	private function performUpdate( Title $title, Skin $skin, &$sidebar ) {
 
 		$link = Infolink::encodeParameters(
 			[
@@ -76,7 +75,7 @@ class SidebarBeforeOutput implements HookListener {
 		);
 
 		$sidebar["TOOLBOX"][] = [
-			'text' => wfMessage( 'smw_browselink' )->text(),
+			'text' => $skin->msg( 'smw_browselink' )->text(),
 			'href' => SpecialPage::getTitleFor( 'Browse', ':' . $link )->getLocalUrl(),
 			'id'   => 't-smwbrowselink',
 			'rel'  => 'search'
