+++ tests/phpunit/MediaWiki/Hooks/SidebarBeforeOutputTest.php
@@ -35,4 +35,155 @@ class SidebarBeforeOutputTest extends \PHPUnit_Framework_TestCase {
 		);
 	}
 
+	/**
+	 * @dataProvider skinTemplateDataProvider
+	 */
+	public function testProcess( $setup, $expected ) {
+
+		$this->namespaceExaminer->expects( $this->any() )
+			->method( 'isSemanticEnabled' )
+			->will( $this->returnValue( $setup['settings']['isEnabledNamespace'] ) );
+
+		$sidebar = [];
+
+		$instance = new SidebarBeforeOutput(
+			$this->namespaceExaminer
+		);
+
+		$instance->setOptions(
+			[
+				'smwgBrowseFeatures' => $setup['settings']['smwgBrowseFeatures']
+			]
+		);
+
+		$this->assertTrue(
+			$instance->process( $setup['skin'], $sidebar )
+		);
+
+		if ( $expected['count'] == 0 ) {
+			$this->assertEmpty( $sidebar );
+		} else {
+			$this->assertCount(
+				$expected['count'],
+				$sidebar['TOOLBOX']
+			);
+		}
+	}
+
+	public function skinTemplateDataProvider() {
+
+		#0 Standard title
+		$settings = [
+			'isEnabledNamespace' => true,
+			'smwgBrowseFeatures' => SMW_BROWSE_TLINK
+		];
+
+		$provider[] = [
+			[
+				'skin' => $this->newSkinStub( true ),
+				'settings' => $settings
+			],
+			[ 'count' => 1 ],
+		];
+
+		#1 isArticle = false
+		$provider[] = [
+			[
+				'skin' => $this->newSkinStub( false ),
+				'settings' => $settings
+			],
+			[ 'count' => 0 ],
+		];
+
+		$settings = [
+			'isEnabledNamespace' => true,
+			'smwgBrowseFeatures' => SMW_BROWSE_NONE
+		];
+
+		$provider[] = [
+			[
+				'skin' => $this->newSkinStub( true ),
+				'settings' => $settings
+			],
+			[ 'count' => 0 ],
+		];
+
+		#3 smwgNamespacesWithSemanticLinks = false
+
+		$settings = [
+			'isEnabledNamespace' => false,
+			'smwgBrowseFeatures' => SMW_BROWSE_TLINK
+		];
+
+		$provider[] = [
+			[
+				'skin' => $this->newSkinStub( true ),
+				'settings' => $settings
+			],
+			[ 'count' => 0 ],
+		];
+
+		#4 Special page
+		$settings = [
+			'isEnabledNamespace' => true,
+			'smwgBrowseFeatures' => SMW_BROWSE_TLINK
+		];
+
+		$title = MockTitle::buildMock( __METHOD__ );
+
+		$title->expects( $this->atLeastOnce() )
+			->method( 'isSpecialPage' )
+			->will( $this->returnValue( true ) );
+
+		$skin = $this->getMockBuilder( '\Skin' )
+			->disableOriginalConstructor()
+			->getMock();
+
+		$skin->expects( $this->atLeastOnce() )
+			->method( 'getTitle' )
+			->will( $this->returnValue( $title ) );
+
+		$provider[] = [
+			[
+				'skin' => $skin,
+				'settings' => $settings
+			],
+			[ 'count' => 0 ],
+		];
+
+		return $provider;
+	}
+
+	private function newSkinStub( bool $isArticle ) {
+
+		$message = $this->getMockBuilder( '\Message' )
+			->disableOriginalConstructor()
+			->getMock();
+
+		$output = $this->getMockBuilder( '\OutputPage' )
+			->disableOriginalConstructor()
+			->getMock();
+
+		$output->expects( $this->atLeastOnce() )
+			->method( 'isArticle' )
+			->willReturn( $isArticle );
+
+		$skin = $this->getMockBuilder( '\Skin' )
+			->disableOriginalConstructor()
+			->getMock();
+
+		$skin->expects( $this->atLeastOnce() )
+			->method( 'getTitle' )
+			->will( $this->returnValue( Title::newFromText( __METHOD__ ) ) );
+
+		$skin->expects( $this->any() )
+			->method( 'msg' )
+			->will( $this->returnValue( $message ) );
+
+		$skin->expects( $this->any() )
+			->method( 'getOutput' )
+			->will( $this->returnValue( $output ) );
+
+		return $skin;
+	}
 }
